Refactor ActionStatus to libs.types for improved organization and clarity
