+++ integrations/discord/events/comparison.py
@@ -12,9 +12,9 @@ from discord.channel import TextChannel
 
 import libs.global_value as g
 from libs.data import modify, search
-from libs.domain.datamodels import ActionStatus, CommandType, ComparisonResults
+from libs.domain.datamodels import CommandType, ComparisonResults
 from libs.domain.score import GameResult
-from libs.types import RemarkDict, StyleOptions
+from libs.types import ActionStatus, RemarkDict, StyleOptions
 from libs.utils import formatter, validator
 from libs.utils.timekit import ExtendedDatetime as ExtDt
 from libs.utils.timekit import Format
