+++ integrations/slack/functions.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Any, cast
 import libs.global_value as g
 from integrations.base.interface import FunctionsInterface
 from libs.data import lookup
-from libs.domain.datamodels import ActionStatus
+from libs.types import ActionStatus
 from libs.utils import validator
 from libs.utils.timekit import Delimiter, Format
 from libs.utils.timekit import ExtendedDatetime as ExtDt
