+++ libs/types.py
@@ -36,6 +36,17 @@ class MessageStatus(StrEnum):
     """未定義状態"""
 
 
+class ActionStatus(StrEnum):
+    """DBに対する操作"""
+
+    CHANGE = "change"
+    """insert/updateが実行された"""
+    DELETE = "delete"
+    """deleteが実行された"""
+    NOTHING = "nothing"
+    """何もしてない"""
+
+
 class ServiceType(StrEnum):
     """連携先サービス"""
